rename to LspDoubleClick, remove count and move to tooling
